mkinitrd: Use --drivers instead of --add-drivers
dracut relies on dracut-install to install files and modules. It passes
`-c` option to dracut-install if `--add-drivers` is specified. The flag
ensures that specified modules exist but it misbehaves for built-in modules
and aliases. dracut-install doesn't consider specified modules as a
built-in and thinks that the module is missing and stops execution with
an error, preventing installation of all external modules following
built-in ones. That leads to broken initrd.

The issue can be workarounded by switching to `--drivers` that only
installs external modules but doesn't fail if it's actually built-in.
But it wouldn't produce an error if the module neither external nor
built-in. So, the approach has to be complemented with a homegrown
validation of module presence.

Three files are searched in order to make sure that all specified kernel
modules can be included in initramfs:

  1) modules.builtin is searched for modules built into kernel
  2) modules.builtin.modinfo is searched for aliased built-in modules
  3) modules.dep is searched for out-of-tree modules

If none of the files contain a required kernel module, then it is
assumed the kernel module is missing and an error is thrown.